+++ src/main/java/ladder/domain/Results.java
@@ -1,16 +1,14 @@
 package ladder.domain;
 
-import java.util.Arrays;
 import java.util.List;
 
-import static ladder.utils.StringUtils.RESULT_REGEX;
 
 public class Results {
 
     private List<String> results;
 
-    public static Results of(String results) {
-        return new Results(Arrays.asList(results.split(RESULT_REGEX)));
+    public static Results of(String[] results) {
+        return new Results(List.of(results));
     }
 
     public Results(List<String> results) {
