--- src/main/java/ladder/utils/StringUtils.java
@@ -1,20 +1,0 @@
-package ladder.utils;
-
-public class StringUtils {
-
-    public static final String ERR_MSG_RESULTS_NUMBER = "실행 결과의 갯수는 참여하는 사람의 수와 동일해야 합니다.";
-    public static final String ERR_MSG_CANDIDATE_NAME = "참여한 사람의 이름을 입력해야 합니다.";
-    public static final String STR_ALL = "all";
-
-    public static final String MSG_INPUT_CANDIDATE_NAMES = "참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)";
-    public static final String MSG_INPUT_MAX_LADDER_HEIGHT = "최대 사다리 높이는 몇 개인가요?";
-    public static final String MSG_INPUT_EXECUTION_RESULTS = "실행 결과를 입력하세요. (결과는 쉼표(,)로 구분하세요)";
-    public static final String MSG_INPUT_RESULT_CANDIDATE_NAME = "결과를 보고싶은 사람은?";
-
-    public static final String MSG_LADDER_RESULT = "사다리 결과\n";
-    public static final String MSG_EXECUTION_RESULT = "실행 결과";
-    public static final String LINE_EXIST = "|-----";
-    public static final String LINE_NO_EXIST = "|     ";
-    public static final String VERTICAL_LINE = "|";
-    public static final String RESULT_REGEX = ",";
-}
