+++ src/main/java/ladder/view/ResultView.java
@@ -6,9 +6,15 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import static ladder.utils.StringUtils.*;
-
 public class ResultView {
+
+    public static final String MSG_LADDER_RESULT = "사다리 결과\n";
+    public static final String MSG_EXECUTION_RESULT = "실행 결과";
+    public static final String LINE_EXIST = "|-----";
+    public static final String LINE_NO_EXIST = "|     ";
+    public static final String VERTICAL_LINE = "|";
+    public static final String RESULT_REGEX = ",";
+    public static final String ALL_RESULT_REGEX = " : ";
 
     public static void printResult(Names names, Ladder ladder) {
         System.out.println(MSG_LADDER_RESULT);
@@ -62,9 +68,10 @@ public class ResultView {
         System.out.println(result);
         System.out.println();
     }
+
     public static void printLadderResultAll(Map<Name, String> ladderResult) {
         System.out.println(MSG_EXECUTION_RESULT);
-        ladderResult.forEach((name, result) -> System.out.println(name.getName() + RESULT_REGIX + result));
+        ladderResult.forEach((name, result) -> System.out.println(name.getName() + ALL_RESULT_REGEX + result));
         System.out.println();
     }
