+++ src/test/java/ladder/domain/LadderTest.java
@@ -3,9 +3,6 @@ package ladder.domain;
 import ladder.utils.LadderUtils;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -20,18 +17,20 @@ class LadderTest {
     @Test
     public void 사다리_결과_인덱스() {
         Ladder ladder = LadderUtils.createLadder();
-        assertThat(ladder.getLadderEndIdx(0)).isEqualTo(2);
-        assertThat(ladder.getLadderEndIdx(1)).isEqualTo(3);
-        assertThat(ladder.getLadderEndIdx(2)).isEqualTo(0);
-        assertThat(ladder.getLadderEndIdx(3)).isEqualTo(1);
+        assertAll(
+                () -> assertThat(ladder.getLadderEndIdx(0)).isEqualTo(2),
+                () -> assertThat(ladder.getLadderEndIdx(1)).isEqualTo(3),
+                () -> assertThat(ladder.getLadderEndIdx(2)).isEqualTo(0),
+                () -> assertThat(ladder.getLadderEndIdx(3)).isEqualTo(1));
     }
 
     @Test
     public void 사다리_결과_테스트2() {
         Ladder ladder = LadderUtils.createLadder2();
-        assertThat(ladder.getLadderEndIdx(0)).isEqualTo(2);
-        assertThat(ladder.getLadderEndIdx(1)).isEqualTo(1);
-        assertThat(ladder.getLadderEndIdx(2)).isEqualTo(0);
-        assertThat(ladder.getLadderEndIdx(3)).isEqualTo(3);
+        assertAll(
+                () -> assertThat(ladder.getLadderEndIdx(0)).isEqualTo(2),
+                () -> assertThat(ladder.getLadderEndIdx(1)).isEqualTo(1),
+                () -> assertThat(ladder.getLadderEndIdx(2)).isEqualTo(0),
+                () -> assertThat(ladder.getLadderEndIdx(3)).isEqualTo(3));
     }
 }
